Move too_many_lines to its own module
